refactor: consolidate search functionality into block and channel modules
