+++ src/config/apollo.ts
@@ -5,12 +5,11 @@ import { Driver } from "neo4j-driver";
 import { authResolvers } from "../resolvers/auth";
 import { blockResolvers } from "../resolvers/block";
 import { channelResolvers } from "../resolvers/channel";
-import { searchResolvers } from "../resolvers/search";
 import { verifyToken } from "../services/jwt";
 import { blockTypeDefs } from "../types/block";
 import { channelTypesDefs } from "../types/channel";
+import { commonTypeDefs } from "../types/common";
 import { Context } from "../types/context";
-import { searchTypeDefs } from "../types/search";
 import { userTypeDefs } from "../types/user";
 
 const baseTypeDefs = `#graphql
@@ -27,17 +26,12 @@ export const createApolloServer = () => {
   return new ApolloServer<Context>({
     typeDefs: [
       baseTypeDefs,
+      commonTypeDefs,
       userTypeDefs,
       channelTypesDefs,
       blockTypeDefs,
-      searchTypeDefs,
     ],
-    resolvers: [
-      authResolvers,
-      channelResolvers,
-      blockResolvers,
-      searchResolvers,
-    ],
+    resolvers: [authResolvers, channelResolvers, blockResolvers],
   });
 };
