+++ src/services/block.ts
@@ -6,9 +6,11 @@ import {
   BlockConnection,
   BlocksByChannelInput,
   BlocksByChannelResult,
+  BlockSearchResult,
   CreateBlockInput,
   UpdateBlockInput,
 } from "../types/block";
+import { SearchPaginationInput } from "../types/common";
 import { UrlService } from "./url";
 
 interface CreateBlockData extends CreateBlockInput {
@@ -84,6 +86,31 @@ export class BlockService {
         tx.run(blockQueries.findBlocksByChannelId, {
           channelId: input.channelId,
           query: input.query ? `(?i).*${input.query}.*` : null,
+          limit: int(input.limit ?? 10),
+          offset: int(input.offset ?? 0),
+        })
+      );
+
+      const blocks = result.records.map((record) => record.get("block"));
+      const totalCount = result.records[0]?.get("totalCount")?.toNumber() || 0;
+      const hasMore = blocks.length + (input.offset ?? 0) < totalCount;
+
+      return {
+        blocks,
+        totalCount,
+        hasMore,
+      };
+    } finally {
+      await session.close();
+    }
+  }
+
+  async searchBlocks(input: SearchPaginationInput): Promise<BlockSearchResult> {
+    const session = this.driver.session();
+    try {
+      const result = await session.executeRead((tx) =>
+        tx.run(blockQueries.searchBlocks, {
+          query: `(?i).*${input.query}.*`,
           limit: int(input.limit ?? 10),
           offset: int(input.offset ?? 0),
         })
